almost done with rename-zipentry
