+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+- 12/12/2023
+    - `ZipEntryBase` Type:
+        - Renamed Property `EntryName` to `Name`.
+        - Renamed Property `EntryRelativePath` to `RelativePath`.
+        - Renamed Property `EntryType` to `Type`.
+        - Renamed Method `RemoveEntry()` to `Remove()`.
+    - `ZipEntryFile` Type:
+        - Added Property `Extension`.
+    - Added Cmdlet `Rename-ZipEntry`.
+
